backend/requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.3.6
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.24
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
